Added plugins and properly implemented pyc.
I wanted a way to get the plugins into the Dockerfile without
ssh-ing into the Pod and performing a git clone. I chose to
bypass this by cloning the plugins I want into this repo as
submodules. This will allow me to easily change their feature
versions in the future. I must proceed with caution if I want
to remove any in the future, however.

I also realized the pycache precompiler wasn't implemented
properly, so I did so.