LIBILS-894. Remove Deployment version tracking
I've removed the deployment tracking. The only thing that needed
this was plugin updates, which was a logic issue anyways.

I had to completely redo the plugin updating system. I now update
a plugin conditionally on whether the incoming plugin is newer
based on the git versions. This also required adding plugins not
as submodules, but cloning them individually.

https://umd-dit.atlassian.net/browse/LIBILS-894