LIBILS-900. Split Nginx and Janeway containers
I found out that, for containers, the best practice is to have
processes running in seperate containers. I've removed nginx
from the Janeway dockerfile and am now using a seperate nginx
image.

As part of this, I also switched Janeway from communicating on
a socket to instead use http. This will remove the need for
a volume mount in kubernetes.

I had to greatly modify the docker-compose configuration to
implement this new deployment. Nginx must wait for Janeway
to be ready. This means using wait-for-dependencies twice.

https://umd-dit.atlassian.net/browse/LIBILS-900