LIBILS-1012. Correct Journal Auto-Installation.
Attempt to correct journal auto-installation. This wasn't working
before because the JANEWAY_JOURNAL_DOMAIN variable was not
recieving the URL for install.

While I was looking at this, I also noticed that the two seperate
variables for Domains and Domain Schemes were unneccesary because
Python includes an easy way to parse this value from the full
URL. So I removed the 'SCHEMES' environment variables and also
replaced the DOMAIN variabes with URL variables.

The user is meant to fill in the URL variables, and the DOMAIN
variables (for journals and press) are auto-generated from the
netloc of the URL variables. It seems the janeway DOMAIN environment
variables need the netloc.

https://umd-dit.atlassian.net/browse/LIBILS-1012