LIBILS-1012. Fix Auto-Installed Journal Domain Feature.
Janeway, during installation, creates a journal. There's a feature
where you can specify a domain for that journal during install.
This required duplicating the domain already specified in the
list of journal domains, so I implemented a variable that you
can set to select a journal domain to be used via index. This
variable only applies during install.

I also refacored the Makefile so that you can use a Makefile.local
to override / set environment variables. This makes it so you
don't have to modify source-controlled files to configure your
default install.

https://umd-dit.atlassian.net/browse/LIBILS-1012